+++ packages/core/tests/js/object/property-node/input.ts
@@ -1,0 +1,4 @@
+const test = {
+	/** a comment */
+	foo: 1
+};